added core library that is consumed by main and lite sdk and added BasePaymentSheetLauncher abstract class
