+++ gradle.properties
@@ -30,4 +30,5 @@ newArchEnabled=false
 hermesEnabled=true
 version=1.1.7
 liteVersion=1.1.4
+coreVersion=1.1.4
 extVersion=0.0.7
