+++ hyperswitch-sdk-android-lite/src/main/java/io/hyperswitch/DefaultPaymentSessionLauncherLite.kt
@@ -1,42 +1,30 @@
-package io.hyperswitch.paymentsession
+package io.hyperswitch
 
 import android.app.Activity
 import android.os.Bundle
-import io.hyperswitch.PaymentConfiguration
 import io.hyperswitch.lite.WebViewUtils
+import io.hyperswitch.paymentsession.BasePaymentSessionLauncher
+import io.hyperswitch.paymentsession.PaymentSessionHandler
+import io.hyperswitch.paymentsession.PaymentSheetCallbackManager
+import io.hyperswitch.paymentsession.SDKInterface
 import io.hyperswitch.paymentsheet.PaymentSheet
 import io.hyperswitch.paymentsheet.PaymentSheetResult
 
 open class DefaultPaymentSessionLauncherLite(
-    private val activity: Activity,
+    activity: Activity,
     publishableKey: String?,
     customBackendUrl: String?,
     customLogUrl: String?,
     customParams: Bundle?,
     private val webViewUtils: SDKInterface = WebViewUtils(activity)
-) : PaymentSessionLauncher {
+) : BasePaymentSessionLauncher(
+    activity,
+    publishableKey,
+    customBackendUrl,
+    customLogUrl,
+    customParams
+) {
 
-    protected var paymentIntentClientSecret: String? = null
-
-    init {
-        if (publishableKey != null) {
-            PaymentConfiguration.init(
-                activity.applicationContext,
-                publishableKey,
-                "",
-                customBackendUrl,
-                customLogUrl,
-                customParams
-            )
-        }
-    }
-
-    // Method to initialize the payment session
-    override fun initPaymentSession(paymentIntentClientSecret: String) {
-        this.paymentIntentClientSecret = paymentIntentClientSecret
-    }
-
-    // Method to present the payment sheet with configuration
     override fun presentPaymentSheet(
         configuration: PaymentSheet.Configuration?,
         resultCallback: (PaymentSheetResult) -> Unit
@@ -53,7 +41,6 @@ open class DefaultPaymentSessionLauncherLite(
         webViewUtils.presentSheet(configurationMap)
     }
 
-    // Get customer saved payment methods
     override fun getCustomerSavedPaymentMethods(
         savedPaymentMethodCallback: ((PaymentSessionHandler) -> Unit)
     ) {
